+++ crates/miden-lib/src/note/utils.rs
@@ -3,7 +3,7 @@ use miden_objects::{
     account::AccountId,
     asset::Asset,
     block::BlockNumber,
-    note::{NoteExecutionMode, NoteInputs, NoteRecipient, NoteTag, NoteType},
+    note::{NoteExecutionMode, NotePayload, NoteRecipient, NoteTag, NoteType},
 };
 
 use super::well_known_note::WellKnownNote;
@@ -17,9 +17,9 @@ pub fn build_p2id_recipient(
     serial_num: Word,
 ) -> Result<NoteRecipient, NoteError> {
     let note_script = WellKnownNote::P2ID.script();
-    let note_inputs = NoteInputs::new(vec![target.suffix(), target.prefix().as_felt()])?;
+    let note_payload = NotePayload::new(vec![target.suffix(), target.prefix().as_felt()])?;
 
-    Ok(NoteRecipient::new(serial_num, note_script, note_inputs))
+    Ok(NoteRecipient::new(serial_num, note_script, note_payload))
 }
 
 /// Creates a [NoteRecipient] for the P2IDE note.
@@ -37,14 +37,14 @@ pub fn build_p2ide_recipient(
     let reclaim_height_u32 = reclaim_block_height.map_or(0, |bn| bn.as_u32());
     let timelock_height_u32 = timelock_block_height.map_or(0, |bn| bn.as_u32());
 
-    let note_inputs = NoteInputs::new(vec![
+    let note_payload = NotePayload::new(vec![
         target.suffix(),
         target.prefix().into(),
         Felt::new(reclaim_height_u32 as u64),
         Felt::new(timelock_height_u32 as u64),
     ])?;
 
-    Ok(NoteRecipient::new(serial_num, note_script, note_inputs))
+    Ok(NoteRecipient::new(serial_num, note_script, note_payload))
 }
 
 /// Returns a note tag for a swap note with the specified parameters.