+++ crates/miden-objects/src/note/recipient.rs
@@ -4,7 +4,7 @@ use core::fmt::Debug;
 use miden_crypto::Felt;
 
 use super::{
-    ByteReader, ByteWriter, Deserializable, DeserializationError, Hasher, NoteInputs, NoteScript,
+    ByteReader, ByteWriter, Deserializable, DeserializationError, Hasher, NotePayload, NoteScript,
     Serializable, Word,
 };
 
@@ -22,12 +22,12 @@ use super::{
 pub struct NoteRecipient {
     serial_num: Word,
     script: NoteScript,
-    inputs: NoteInputs,
+    inputs: NotePayload,
     digest: Word,
 }
 
 impl NoteRecipient {
-    pub fn new(serial_num: Word, script: NoteScript, inputs: NoteInputs) -> Self {
+    pub fn new(serial_num: Word, script: NoteScript, inputs: NotePayload) -> Self {
         let digest = compute_recipient_digest(serial_num, &script, &inputs);
         Self { serial_num, script, inputs, digest }
     }
@@ -46,7 +46,7 @@ impl NoteRecipient {
     }
 
     /// The recipient's inputs which customizes the script's behavior.
-    pub fn inputs(&self) -> &NoteInputs {
+    pub fn inputs(&self) -> &NotePayload {
         &self.inputs
     }
 
@@ -62,8 +62,8 @@ impl NoteRecipient {
     /// The format is `inputs_length || INPUTS_COMMITMENT || SCRIPT_ROOT || SERIAL_NUMBER`
     ///
     /// Where:
-    /// - inputs_length is the length of the note inputs
-    /// - INPUTS_COMMITMENT is the commitment of the note inputs
+    /// - inputs_length is the length of the note payload
+    /// - INPUTS_COMMITMENT is the commitment of the note payload
     /// - SCRIPT_ROOT is the commitment of the note script (i.e., the script's MAST root)
     /// - SERIAL_NUMBER is the recipient's serial number
     pub fn format_for_advice(&self) -> Vec<Felt> {
@@ -76,7 +76,7 @@ impl NoteRecipient {
     }
 }
 
-fn compute_recipient_digest(serial_num: Word, script: &NoteScript, inputs: &NoteInputs) -> Word {
+fn compute_recipient_digest(serial_num: Word, script: &NoteScript, inputs: &NotePayload) -> Word {
     let serial_num_hash = Hasher::merge(&[serial_num, Word::empty()]);
     let merge_script = Hasher::merge(&[serial_num_hash, script.root()]);
     Hasher::merge(&[merge_script, inputs.commitment()])
@@ -106,7 +106,7 @@ impl Serializable for NoteRecipient {
 impl Deserializable for NoteRecipient {
     fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
         let script = NoteScript::read_from(source)?;
-        let inputs = NoteInputs::read_from(source)?;
+        let inputs = NotePayload::read_from(source)?;
         let serial_num = Word::read_from(source)?;
 
         Ok(Self::new(serial_num, script, inputs))