+++ crates/miden-tx/src/host/note_builder.rs
@@ -2,7 +2,7 @@ use alloc::{boxed::Box, vec::Vec};
 
 use miden_objects::{
     asset::Asset,
-    note::{Note, NoteAssets, NoteInputs, NoteMetadata, NoteRecipient, NoteScript, PartialNote},
+    note::{Note, NoteAssets, NoteMetadata, NotePayload, NoteRecipient, NoteScript, PartialNote},
 };
 use vm_processor::AdviceProvider;
 
@@ -67,7 +67,7 @@ impl OutputNoteBuilder {
 
             let inputs_data = adv_provider.get_mapped_values(&inputs_commitment);
             let inputs = match inputs_data {
-                Err(_) => NoteInputs::default(),
+                Err(_) => NotePayload::default(),
                 Ok(inputs) => {
                     let num_inputs = data[0].as_int() as usize;
 
@@ -78,19 +78,19 @@ impl OutputNoteBuilder {
                     // will be discarded below, and later their contents will be validated by
                     // computing the commitment and checking against the expected value.
                     if num_inputs > inputs.len() {
-                        return Err(TransactionKernelError::TooFewElementsForNoteInputs {
+                        return Err(TransactionKernelError::TooFewElementsForNotePayload {
                             specified: num_inputs as u64,
                             actual: inputs.len() as u64,
                         });
                     }
 
-                    NoteInputs::new(inputs[0..num_inputs].to_vec())
-                        .map_err(TransactionKernelError::MalformedNoteInputs)?
+                    NotePayload::new(inputs[0..num_inputs].to_vec())
+                        .map_err(TransactionKernelError::MalformedNotePayload)?
                 },
             };
 
             if inputs.commitment() != inputs_commitment {
-                return Err(TransactionKernelError::InvalidNoteInputs {
+                return Err(TransactionKernelError::InvalidNotePayload {
                     expected: inputs_commitment,
                     actual: inputs.commitment(),
                 });