+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.11.0 (TBD)
 
+- [BREAKING] Rename `NoteInputs` to `NotePayload`.
+
 ### Features
 
 - Added arguments to the auth procedure ([#1501](https://github.com/0xMiden/miden-base/pull/1501)).